extracted out existing notification logic into it's own service class
