+++ Shared/Views/MobSessionView/MobSessionView.swift
@@ -58,7 +58,7 @@ struct MobSessionView: View {
             .sheet(isPresented: $showingInfoSheet) {
                 OnboardingView(firstTime: $showingInfoSheet)
             }
-            .onAppear(perform: vm.requestNotificationPermission)
+            .onAppear(perform: vm.localNotificationService.requestNotificationPermission)
             .onReceive(NotificationCenter.default.publisher(for: UIApplication.didEnterBackgroundNotification)) { _ in
                 vm.movedToBackground()
             }
